Question and some refactoring

+++ resources/js/mixins/DeleteActionMixin.vue
@@ -15,9 +15,14 @@ export default {
     },
     methods: {
         onDeleteClick: function(row) {
-            axios.delete(`${this.endpoint}/${row.id}`).then(() => {
-                this.reloadPager();
-            });
+            alertify.confirm(
+                "Are you sure you want to delete this entry?",
+                () => {
+                    axios.delete(`${this.endpoint}/${row.id}`).then(() => {
+                        this.reloadPager();
+                    });
+                }
+            );
         }
     }
 };
